Move hook-implementor traits to `common` and client traits to `hooks`
