+++ src/hooks/mod.rs
@@ -5,7 +5,12 @@
 //!
 //! [`imgui`]: https://docs.rs/imgui/0.8.0/imgui/
 
-pub(crate) mod common;
+use imgui::{Context, Io, Ui};
+use windows::Win32::Foundation::{HWND, LPARAM, WPARAM};
+
+pub use crate::hooks::common::Hooks;
+
+pub mod common;
 #[cfg(feature = "dx11")]
 pub mod dx11;
 #[cfg(feature = "dx12")]
@@ -15,32 +20,39 @@ pub mod dx9;
 #[cfg(feature = "opengl3")]
 pub mod opengl3;
 
-pub use common::{ImguiRenderLoop, ImguiRenderLoopFlags};
-
-/// Generic trait for platform-specific hooks.
-pub trait Hooks {
-    fn from_render_loop<T>(t: T) -> Box<Self>
-    where
-        Self: Sized,
-        T: ImguiRenderLoop + Send + Sync + 'static;
-
-    /// Find the hook target functions addresses, initialize the data, create
-    /// and enable the hooks.
-    ///
-    /// # Safety
-    ///
-    /// Is most definitely UB.
-    unsafe fn hook(&self);
-
-    /// Cleanup global data and disable the hooks.
-    ///
-    /// # Safety
-    ///
-    /// Is most definitely UB.
-    unsafe fn unhook(&mut self);
+/// Holds information useful to the render loop which can't be retrieved from
+/// `imgui::Ui`.
+pub struct ImguiRenderLoopFlags {
+    /// Whether the hooked program's window is currently focused.
+    pub focused: bool,
 }
 
-pub fn initialize() {}
+/// Implement your `imgui` rendering logic via this trait.
+pub trait ImguiRenderLoop {
+    /// Called once at the first occurrence of the hook. Implement this to
+    /// initialize your data.
+    fn initialize(&mut self, _ctx: &mut Context) {}
+
+    /// Called every frame. Use the provided `ui` object to build your UI.
+    fn render(&mut self, ui: &mut Ui, flags: &ImguiRenderLoopFlags);
+
+    /// Called during the window procedure.
+    fn on_wnd_proc(&self, _hwnd: HWND, _umsg: u32, _wparam: WPARAM, _lparam: LPARAM) {}
+
+    /// If this function returns true, the WndProc function will not call the
+    /// procedure of the parent window.
+    fn should_block_messages(&self, _io: &Io) -> bool {
+        false
+    }
+
+    fn into_hook<T>(self) -> Box<T>
+    where
+        T: Hooks,
+        Self: Send + Sync + Sized + 'static,
+    {
+        T::from_render_loop(self)
+    }
+}
 
 #[inline]
 fn loword(l: u32) -> u16 {
